+++ src/app/components/link.tsx
@@ -3,12 +3,20 @@ import React from 'react'
 export function Link({
   href,
   children,
+  className,
+  target,
+  rel,
+  ...rest
 }: {
   href: string
   children: React.ReactNode
+  className?: string
+  target?: string
+  rel?: string
+  [key: string]: any
 }) {
   return (
-    <a className="text-500" href={href}>
+    <a className={className || "text-500"} href={href} target={target} rel={rel} {...rest}>
       {children}
     </a>
   )